Update default widget proxy
