+++ components/widgets/fires/fires-alerts-cumulative/index.js
@@ -2,11 +2,10 @@ import { all, spread } from 'axios';
 import uniq from 'lodash/uniq';
 
 import { fetchVIIRSAlerts, fetchVIIRSLatest } from 'services/analysis-cached';
-import burnedAreaCumulative from 'components/widgets/fires/burned-area-cumulative';
 
 import getWidgetProps from './selectors';
 
-const defaultConfig = {
+export default {
   widget: 'firesAlertsCumulative',
   title: 'Cumulative Fire Alerts in {location}',
   large: true,
@@ -312,20 +311,4 @@ const defaultConfig = {
     ),
   getDataURL: (params) => [fetchVIIRSAlerts({ ...params, download: true })],
   getWidgetProps,
-};
-
-export default {
-  widget: 'firesAlertsCumulative',
-  proxy: true,
-  refetchKeys: ['dataset'],
-  getWidget: (widgetSettings) => {
-    // called when settings changes
-    if (!widgetSettings || !widgetSettings.dataset) {
-      return defaultConfig;
-    }
-    if (widgetSettings.dataset === 'modis_burned_area') {
-      return burnedAreaCumulative;
-    }
-    return defaultConfig;
-  },
 };
